+++ lib/has_states/configuration/state_type_configuration.rb
@@ -4,12 +4,13 @@ module HasStates
   class Configuration
     class StateTypeConfiguration
       attr_reader :name
-      attr_accessor :statuses, :limit
+      attr_accessor :statuses, :limit, :metadata_schema
 
       def initialize(name)
         @name = name
         @statuses = []
         @limit = nil
+        @metadata_schema = nil
       end
     end
   end